+++ examples/read-logs.example.ts
@@ -1,5 +1,8 @@
 import { LogSeverity } from "../src/domain/enums/logSeverity.enum.js";
-import { logger } from "../src/logger.js";
+
+import { createLogger } from "../src/index.js";
+
+const logger = createLogger();
 
 const logs = await logger.getLogsBySeverity(LogSeverity.error);