+++ src/domain/repositories/log.repository.ts
@@ -2,7 +2,7 @@ import type { LogEntity } from "../entities/log.entity.js";
 import type { LogSeverity } from "../enums/logSeverity.enum.js";
 
 export abstract class LogRepository {
-    abstract readLogs(severity: LogSeverity): Promise<LogEntity[]>;
+    abstract readLogs(severity?: LogSeverity): Promise<LogEntity[]>;
     abstract saveLog(log: LogEntity): Promise<void>;
 
     // TODO implementar los siguientes metodos